More tweaks. Hopefully vim/bash settings will follow me through ssh

+++ install.sh
@@ -1,22 +1,30 @@
 #!/bin/bash
 
+if [ -f ~/.vim/joerc ]; then
+	echo "Server configured with joerc"
+fi
+
 goodinput="no"
 
-while [ "$goodinput" == "no" ]; do
+if [ -f /etc/redhad-release ]; then
+	sudo yum -y install git vim ctags	
+else
+	while [ "$goodinput" == "no" ]; do
 
-	echo "1: Ubuntu"
-	echo "2: CentOS"
-	read -p "Choose your OS: " os
+		echo "1: Ubuntu"
+		echo "2: CentOS"
+		read -p "Choose your OS: " os
 
-	if [ "$os" == "1"  ]; then
-		sudo apt-get install git vim exuberant-ctags
-		goodinput="yes"
-	fi
-	if [ "$os" ==  "2" ]; then
-		sudo yum -y install git vim ctags
-		goodinput="yes"
-	fi
-done 
+		if [ "$os" == "1"  ]; then
+			sudo apt-get install git vim exuberant-ctags
+			goodinput="yes"
+		fi
+		if [ "$os" ==  "2" ]; then
+			sudo yum -y install git vim ctags
+			goodinput="yes"
+		fi
+	done 
+fi
 
 mkdir -p ~/.vim
 cp ./vimfiles.tar.gz ~/.vim/
@@ -24,7 +32,12 @@ cd ~/.vim
 tar -xzvf ./vimfiles.tar.gz
 rm -f vimfiles.tar.gz
 mv ~/.vim/.vimrc ~/
+mv ~/.vim/.bashrc ~/
+mkdir -p ~/joerc
+mv ~/.vim/ssh.sh ~/joerc
+mv ~/.vim/run.sh ~/joerc
 
+touch ~/.vim/joerc
 
 cat << EOF
 =======================================================
